Create separate DefinitionLevelHolder
